requirements.txt
streamlit==1.38.0
openai==1.44.0
pillow==10.4.0
pytest==8.3.3
pytest-mock==3.14.0
black==24.8.0
ruff==0.6.9
